+++ db/structure.sql
@@ -5103,5 +5103,3 @@ INSERT INTO "schema_migrations" (version) VALUES
 ('20220413084748'),
 ('20220504090512'),
 ('20220524130709');
-
-