+++ test/system/registrar_area/invoices/new_invoice_payment_test.rb
@@ -3,6 +3,8 @@ require 'application_system_test_case'
 class NewInvoicePaymentTest < ApplicationSystemTestCase
   def setup
     super
+    eis_response = OpenStruct.new(body: "{\"payment_link\":\"http://link.test\"}")
+    Spy.on_instance_method(EisBilling::AddDeposits, :send_invoice).and_return(eis_response)
 
     @original_vat_prc = Setting.registry_vat_prc
     Setting.registry_vat_prc = 0.2
@@ -22,27 +24,5 @@ class NewInvoicePaymentTest < ApplicationSystemTestCase
     fill_in 'Amount', with: '200.00'
     fill_in 'Description', with: 'My first invoice'
     click_link_or_button 'Add'
-  end
-
-  def test_create_new_SEB_payment
-    create_invoice_and_visit_its_page
-    click_link_or_button 'seb'
-    form = page.find('form')
-    assert_equal('https://www.seb.ee/cgi-bin/dv.sh/ipank.r', form['action'])
-    assert_equal('post', form['method'])
-    assert_equal('240.00', form.find_by_id('VK_AMOUNT', visible: false).value)
-  end
-
-  def test_create_new_Every_Pay_payment
-    create_invoice_and_visit_its_page
-    click_link_or_button 'every_pay'
-    expected_hmac_fields = 'account_id,amount,api_username,callback_url,' +
-                           'customer_url,hmac_fields,nonce,order_reference,timestamp,transaction_type'
-
-    form = page.find('form')
-    assert_equal('https://igw-demo.every-pay.com/transactions/', form['action'])
-    assert_equal('post', form['method'])
-    assert_equal(expected_hmac_fields, form.find_by_id('hmac_fields', visible: false).value)
-    assert_equal('240.00', form.find_by_id('amount', visible: false).value)
   end
 end