+++ test/tasks/check_force_delete_test.rb
@@ -54,6 +54,7 @@ class CheckForceDeleteTaskTest < ActiveSupport::TestCase
       action.call
     end
 
+    run_task
     assert_enqueued_jobs 0
   end